+++ src/pymetropolis/metro_network/bicycle_network/osm.py
@@ -495,6 +495,9 @@ class OpenStreetMapBicycleImportStep(GeoStep, OSMStep):
     - `give_way`: way has a node with `highway=give_way`, in the correct direction.
     - `stop`: way has a node with `highway=stop`, in the correct direction.
     - `traffic_signals`: way has a node with `highway=traffic_signals`, in the correct direction.
+
+    If the [`reindex`](parameters.md#osm_bicycle_importreindex) parameter is set to `true`, the
+    `edge_id` values are instead numerical values running from 1 to the number of edges.
     """
 
     highways = ListParameter(
@@ -508,6 +511,14 @@ class OpenStreetMapBicycleImportStep(GeoStep, OSMStep):
         note=(
             "A list of highway tags with description is available on the "
             "[OpenStreetMap wiki](https://wiki.openstreetmap.org/wiki/Key:highway)."
+        ),
+    )
+    reindex = BoolParameter(
+        "osm_pedestrian_import.reindex",
+        default=False,
+        description=(
+            "If `true`, the edges are re-index from 1 to n. If `false`, edge ids match the "
+            "OpenStreetMap way ids."
         ),
     )
     simulation_area_filter = BoolParameter(
@@ -554,6 +565,7 @@ class OpenStreetMapBicycleImportStep(GeoStep, OSMStep):
             highway_tags=self.highways,
             crs=self.crs,
             filter_polygon=filter_polygon,
+            reindex=self.reindex,
         )
         edges = importer.run()
         self.output["raw_edges"].write(edges)